# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraclyTag Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  master: "master",
  client_admin: "client_admin",
  operator: "operator",
} as const;

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthSession {
  id: number;
  username: string;
  email: string;
  role: Role;
  /** @nullable */
  companyId: number | null;
  /** @nullable */
  companyName: string | null;
}

export interface Company {
  id: number;
  name: string;
  email: string;
  address: string;
  /** @nullable */
  gstin: string | null;
  createdAt: string;
}

export interface CreateCompanyBody {
  name: string;
  email: string;
  address: string;
  /** @nullable */
  gstin?: string | null;
}

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  phone: string | null;
  role: Role;
  /** @nullable */
  companyId: number | null;
  /** @nullable */
  companyName: string | null;
  createdAt: string;
}

export interface CreateUserBody {
  username: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  password: string;
  role: Role;
  /** @nullable */
  companyId?: number | null;
}

export interface Product {
  id: number;
  companyId: number;
  skuId: string;
  name: string;
  skuSize: string;
  marketedBy: string;
  /** @nullable */
  sapDescription: string | null;
  gtin: string;
  mrp: number;
  /** @nullable */
  registrationNo: string | null;
  /** Units per 1st level inner pack */
  l1Size: number;
  /** 1st level packs per 2nd level inner pack */
  l2Size: number;
  /** 2nd level packs per shipper */
  shipperSize: number;
  /** @nullable */
  cautionLogoUrl: string | null;
  /** @nullable */
  productLogoUrl: string | null;
  /** @nullable */
  labelPdfUrl: string | null;
  createdAt: string;
}

export interface CreateProductBody {
  skuId: string;
  name: string;
  skuSize: string;
  marketedBy: string;
  /** @nullable */
  sapDescription?: string | null;
  /** 14 digit GTIN */
  gtin: string;
  mrp: number;
  /** @nullable */
  registrationNo?: string | null;
  l1Size: number;
  l2Size: number;
  shipperSize: number;
  /** @nullable */
  cautionLogoUrl?: string | null;
  /** @nullable */
  productLogoUrl?: string | null;
  /** @nullable */
  labelPdfUrl?: string | null;
}

export interface Location {
  id: number;
  companyId: number;
  locationType: string;
  uniqueName: string;
  locationName: string;
  contactNo: string;
  state: string;
  city: string;
  address: string;
  createdAt: string;
}

export interface CreateLocationBody {
  locationType: string;
  uniqueName: string;
  locationName: string;
  contactNo: string;
  state: string;
  city: string;
  address: string;
}

export interface Batch {
  id: number;
  productId: number;
  productName: string;
  batchNumber: string;
  expiryDate: string;
  createdAt: string;
}

export interface CreateBatchBody {
  productId: number;
  batchNumber: string;
  expiryDate: string;
}

export type CodeLevel = (typeof CodeLevel)[keyof typeof CodeLevel];

export const CodeLevel = {
  unit: "unit",
  l1: "l1",
  l2: "l2",
  shipper: "shipper",
  pallet: "pallet",
} as const;

export interface Code {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  batchId: number | null;
  /** @nullable */
  batchNumber: string | null;
  level: CodeLevel;
  /** Raw GS1 string with FNC1 (ASCII 232) where applicable, no brackets */
  rawString: string;
  /** @nullable */
  serialNumber: string | null;
  /** @nullable */
  ssccCode: string | null;
  mapped: boolean;
  /** @nullable */
  mappedAt: string | null;
  /** @nullable */
  mappedByUserId: number | null;
  /** @nullable */
  mappedByUsername: string | null;
  /** @nullable */
  locationId: number | null;
  /** @nullable */
  locationName: string | null;
  createdAt: string;
}

export interface GenerateCodesBody {
  batchId: number;
  level: CodeLevel;
  /**
   * @minimum 1
   * @maximum 5000
   */
  quantity: number;
}

export interface GenerateCodesResponse {
  generated: number;
  codes: Code[];
}

export interface MapCodeBody {
  locationId: number;
}

export interface LevelCount {
  level: CodeLevel;
  count: number;
}

export interface DashboardSummary {
  totalProducts: number;
  totalBatches: number;
  totalCodes: number;
  totalMapped: number;
  totalUnmapped: number;
  totalLocations: number;
  totalUsers: number;
  totalCompanies: number;
  recentCodes: Code[];
  codesByLevel: LevelCount[];
}

export interface StockRow {
  productId: number;
  productName: string;
  batchId: number;
  batchNumber: string;
  totalCodes: number;
  mapped: number;
  unmapped: number;
  /** @nullable */
  location: string | null;
}

export interface ProductReportRow {
  productId: number;
  productName: string;
  batchId: number;
  batchNumber: string;
  size: string;
  total: number;
  mapped: number;
  unmapped: number;
}

export interface ShipperSummaryRow {
  productId: number;
  productName: string;
  batchId: number;
  batchNumber: string;
  size: string;
  total: number;
  shipperMapped: number;
  shipperUnmapped: number;
}

export interface PalletSummaryRow {
  productId: number;
  productName: string;
  batchId: number;
  batchNumber: string;
  size: string;
  total: number;
  palletMapped: number;
  palletUnmapped: number;
}

export interface MarkedByEntry {
  codeId: number;
  rawString: string;
  level: CodeLevel;
  productName: string;
  /** @nullable */
  batchNumber: string | null;
  /** @nullable */
  locationName: string | null;
  mappedByUsername: string;
  mappedAt: string;
}

export type ListBatchesParams = {
  productId?: number;
};

export type ListCodesParams = {
  level?: CodeLevel;
  batchId?: number;
  productId?: number;
  limit?: number;
};

export type GetStockReportParams = {
  productId?: number;
  from?: string;
  to?: string;
};

export type GetProductReportParams = {
  from?: string;
  to?: string;
};

export type GetShipperSummaryParams = {
  from?: string;
  to?: string;
};

export type GetPalletSummaryParams = {
  from?: string;
  to?: string;
};