# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TraclyTag Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- COMPANIES (master only) ----------
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    delete:
      operationId: deleteCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- USERS ----------
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    delete:
      operationId: deleteLocation
      tags: [locations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- BATCHES ----------
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [batches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    delete:
      operationId: deleteBatch
      tags: [batches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- CODES ----------
  /codes:
    get:
      operationId: listCodes
      tags: [codes]
      parameters:
        - name: level
          in: query
          schema:
            $ref: "#/components/schemas/CodeLevel"
        - name: batchId
          in: query
          schema:
            type: integer
        - name: productId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Code"
    post:
      operationId: generateCodes
      tags: [codes]
      summary: Generate a quantity of GS1 codes for a batch at a given level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCodesBody"
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateCodesResponse"
  /codes/{id}/map:
    post:
      operationId: mapCode
      tags: [codes]
      summary: Mark a code as mapped to a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapCodeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Code"

  # ---------- REPORTS ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Role:
      type: string
      enum: [master, client_admin, operator]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthSession:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
      required: [id, username, email, role, companyId, companyName]
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, address, gstin, createdAt]
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: ["string", "null"]
      required: [name, email, address]
...
    Location:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        locationType:
          type: string
        uniqueName:
          type: string
        locationName:
          type: string
        contactNo:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - locationType
        - uniqueName
        - locationName
        - contactNo
        - state
        - city
        - address
        - createdAt
    CreateLocationBody:
      type: object
      properties:
        locationType:
          type: string
        uniqueName:
          type: string
        locationName:
          type: string
        contactNo:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
      required:
        - locationType
        - uniqueName
        - locationName
        - contactNo
        - state
        - city
        - address
    Batch:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        batchNumber:
          type: string
        expiryDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - productName
        - batchNumber
        - expiryDate
        - createdAt
    CreateBatchBody:
      type: object
      properties:
        productId:
          type: integer
        batchNumber:
          type: string
        expiryDate:
          type: string
          format: date
      required: [productId, batchNumber, expiryDate]
    CodeLevel:
      type: string
      enum: [unit, l1, l2, shipper, pallet]
    Code:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: ["integer", "null"]
        batchNumber:
          type: ["string", "null"]
        level:
          $ref: "#/components/schemas/CodeLevel"
        rawString:
          type: string
          description: Raw GS1 string with FNC1 (ASCII 232) where applicable, no brackets
        serialNumber:
          type: ["string", "null"]
        ssccCode:
          type: ["string", "null"]
        mapped:
          type: boolean
        mappedAt:
          type: ["string", "null"]
          format: date-time
        mappedByUserId:
          type: ["integer", "null"]
        mappedByUsername:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - productName
        - batchId
        - batchNumber
        - level
        - rawString
        - serialNumber
        - ssccCode
        - mapped
        - mappedAt
        - mappedByUserId
        - mappedByUsername
        - locationId
        - locationName
        - createdAt
    GenerateCodesBody:
      type: object
      properties:
        batchId:
          type: integer
        level:
          $ref: "#/components/schemas/CodeLevel"
        quantity:
          type: integer
          minimum: 1
          maximum: 5000
      required: [batchId, level, quantity]
    GenerateCodesResponse:
      type: object
      properties:
        generated:
          type: integer
        codes:
          type: array
          items:
            $ref: "#/components/schemas/Code"
      required: [generated, codes]
    MapCodeBody:
      type: object
      properties:
        locationId:
          type: integer
      required: [locationId]
...