# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraclyTag Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface AuthSession {
  id: number;
  username: string;
  email: string;
  role: Role;
  /** @nullable */
  companyId: number | null;
  /** @nullable */
  companyName: string | null;
}