# lib/api-zod/src/generated/types/createCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraclyTag Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCompanyBody {
  name: string;
  email: string;
  address: string;
  /** @nullable */
  gstin?: string | null;
}