# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TraclyTag Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  master: "master",
  client_admin: "client_admin",
  operator: "operator",
} as const;